before: denominate bit streams in u64 end to end; gate encoded_bits meter-side
Stored streams, cursors, walks, and depths now speak one width: u64,
the denomination the build buffer and the view already carry. Bits::len
returns the u64 live length directly; the freeze and decode doors adopt
any buffer whose padding validates, imposing no bound of their own; the
BitCursor trait's position and read_unary are u64 (the separate u64
position read dissolves into them); BitStack heights, unary runs, tree
depths, flip levels, code lengths, and open-range counts widen across
the walk surface. On 32-bit targets streams are therefore bounded only
by allocatable memory: no usize-derived structural cap survives on the
stored, walk, or build path. Byte indexes stay usize (they index
allocated buffers), and each remaining u64-to-usize seam carries a
checked conversion with its rationale inline.

Where a gamma mantissa names a value wider than the big-integer backend
can hold on the target, both decode paths now reject with the value's
own genre (NotCanonical) at the same width, per-bit and word-parallel
alike.

Owner ruling (Finch, direct): encoded_bits on Party, Version, and Clock
leaves the unconditional public API — every consumer is a meter, a
coverage suite, a fuel denominator, or a boundary pin — and re-homes
behind the existing meter feature (the instrument surface), returning
u64. Consumers routed: laws now implies meter (the representational
laws read the observation), the fuzzfit harness and wasm32-pins guest
enable it explicitly, and rumors' dev-dependency lights it for the
conservation suite, whose invariant is unchanged. The surface roster's
rows are unchanged: the extractor scans source text, so the gated fns
keep their named rows.

The recode mutant exclusion's line pin moves with its expression
(rustfmt reflowed the file); same mutant, same rationale, re-pinned.
Meter parity: all 300 MEASURED rows are byte-identical to the parent
commit's readings.

+++ crates/before-fuelscape/src/sample/tests.rs
@@ -240,7 +240,7 @@ proptest! {
             .expect("every byte size down to 1 has canonical versions");
         prop_assert_eq!(draw.bytes.len(), size);
         let version = Version::decode(&draw.bytes[..]).expect("sampler output must be canonical");
-        prop_assert_eq!(version.encoded_bits(), draw.bits);
+        prop_assert_eq!(version.encoded_bits(), draw.bits as u64);
         prop_assert_eq!(version.encode(), draw.bytes);
     }
 
@@ -257,7 +257,7 @@ proptest! {
             .expect("every byte size down to 1 has canonical parties");
         prop_assert_eq!(draw.bytes.len(), size);
         let party = Party::decode(&draw.bytes[..]).expect("sampler output must be canonical");
-        prop_assert_eq!(party.encoded_bits(), draw.bits);
+        prop_assert_eq!(party.encoded_bits(), draw.bits as u64);
         prop_assert_eq!(party.encode(), draw.bytes);
     }
 }
